[Refactor] Login 구조 재정리(#61)
